[Add] move observer to hook
